Añadir test.c, cambiar librerías a local, quitar variable buffer y MSGERR local en head().
Añadir el archivo test.c que es donde se tiene que ejecutar el main(), y se ha cambiado el include de libreria.h para que se detecte como local poniendo comillas.
También se ha eliminado la variable buffer de la función callFunct porque no se estaba utilizando.
Por último, head() tiene que implementar MSGERR como variable local porque así lo pide la interfaz, se ha cambiado.
